Work on ReadOnly Mode

+++ entity/pyx-entity-server/src/main/java/com/pyx4j/entity/server/EntityServicesImpl.java
@@ -25,8 +25,10 @@ import java.util.Vector;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.xnap.commons.i18n.I18n;
 
 import com.pyx4j.commons.Consts;
+import com.pyx4j.config.server.ServerSideConfiguration;
 import com.pyx4j.entity.rpc.EntityCriteriaByPK;
 import com.pyx4j.entity.rpc.EntitySearchResult;
 import com.pyx4j.entity.rpc.EntityServices;
@@ -36,6 +38,8 @@ import com.pyx4j.entity.server.search.SearchResultIterator;
 import com.pyx4j.entity.shared.IEntity;
 import com.pyx4j.entity.shared.criterion.EntityQueryCriteria;
 import com.pyx4j.entity.shared.criterion.EntitySearchCriteria;
+import com.pyx4j.i18n.shared.I18nFactory;
+import com.pyx4j.rpc.shared.UnRecoverableRuntimeException;
 import com.pyx4j.rpc.shared.VoidSerializable;
 import com.pyx4j.security.shared.SecurityController;
 
@@ -43,10 +47,19 @@ public class EntityServicesImpl {
 
     private static final Logger log = LoggerFactory.getLogger(EntityServicesImpl.class);
 
+    private static I18n i18n = I18nFactory.getI18n();
+
+    public static String applicationReadOnlyMessage() {
+        return i18n.tr("Application is in read-only due to short maintenance.\nPlease try again in one hour");
+    }
+
     public static class SaveImpl implements EntityServices.Save {
 
         @Override
         public IEntity execute(IEntity request) {
+            if (ServerSideConfiguration.instance().datastoreReadOnly()) {
+                throw new UnRecoverableRuntimeException(applicationReadOnlyMessage());
+            }
             if (request.getPrimaryKey() == null) {
                 SecurityController.assertPermission(EntityPermission.permissionCreate(request.getObjectClass()));
             } else {
@@ -61,6 +74,9 @@ public class EntityServicesImpl {
 
         @Override
         public IEntity execute(IEntity request) {
+            if (ServerSideConfiguration.instance().datastoreReadOnly()) {
+                throw new UnRecoverableRuntimeException(applicationReadOnlyMessage());
+            }
             if (request.getPrimaryKey() == null) {
                 SecurityController.assertPermission(EntityPermission.permissionCreate(request.getObjectClass()));
             } else {
@@ -163,6 +179,9 @@ public class EntityServicesImpl {
 
         @Override
         public VoidSerializable execute(IEntity entity) {
+            if (ServerSideConfiguration.instance().datastoreReadOnly()) {
+                throw new UnRecoverableRuntimeException(applicationReadOnlyMessage());
+            }
             SecurityController.assertPermission(new EntityPermission(entity.getObjectClass(), EntityPermission.DELETE));
             PersistenceServicesFactory.getPersistenceService().delete(entity);
             return null;
